clarify import prerequisite for lerna import
On the Github page for this method, this is covered: https://github.com/lerna/lerna/tree/master/commands/import#readme
It's also covered in the troubleshooting doc: https://github.com/lerna/lerna/blob/master/doc/troubleshooting.md#failing-when-git-tree-has-uncommitted-changes

However, that's not where first-timers (like myself) start. I started on the website, and this was very unclear and the error message very unhelpful. In lieu of a better error message from the CLI, this should suffice.